Use 'wget' in Prebuilt-MPR setup instructions
Previously 'curl' was used, but that's not installed by default on Debian/Ubuntu, which can cause issues when the error gets overlooked, or the user just doesn't know what's happening in the error. This changes the setup command to use 'wget' instead (which is installed by default on Debian and Ubuntu), making it less likely for the user to encounter an error.